+++ src/bark_tab/generate_choice_string.py
@@ -1,5 +1,5 @@
 from src.bark_tab.create_voice_string import create_voice_string
-from src.get_speaker_gender import get_speaker_gender
+from src.bark_tab.get_speaker_gender import get_speaker_gender
 
 import gradio as gr
