+++ src/bark_tab/history_to_hash.py
@@ -1,6 +1,6 @@
 import hashlib
 
-from .FullGeneration import FullGeneration
+from src.bark_tab.FullGeneration import FullGeneration
 
 
 def history_to_hash(npz: FullGeneration):
