+++ src/config/config.py
@@ -1,3 +1,3 @@
-from src.load_config import load_config
+from src.config.load_config import load_config
 
 config = load_config()
