+++ src/css/css.py
@@ -1,0 +1,10 @@
+from src.bark_tab.generation_tab_bark import bark_css
+from src.css.material_symbols_css import material_symbols_css
+from src.history_tab.history_css import history_css
+from src.tortoise.generation_tab_tortoise import css_tortoise
+
+full_css = ""
+full_css += material_symbols_css
+full_css += css_tortoise
+full_css += history_css
+full_css += bark_css
