+++ src/css/material_symbols_css.py
