+++ src/extensions_loader/extensions/callback_save_generation_ffmpeg.py
@@ -6,7 +6,6 @@ from models.bark.bark.generation import SAMPLE_RATE
 
 import subprocess
 import ffmpeg
-import numpy as np
 
 import base64
